requirements.txt
streamlit
rembg
Pillow
numpy
scikit-image
onnxruntime
